+++ src/components/IdeaForm/index.tsx
@@ -1,47 +1,7 @@
 import classNames from "classnames";
-import { SyntheticEvent, useCallback, useState } from "react";
+import { SyntheticEvent, useCallback } from "react";
 import AlertError from "src/components/AlertError";
-import { trpc } from "src/utils/trpc";
-
-type UseIdeaFormProps = {
-  onCreate: () => void;
-};
-function useIdeaForm({ onCreate }: UseIdeaFormProps) {
-  const mutation = trpc.useMutation("idea.createIdea");
-  const [description, setDescription] = useState("");
-  const [isLoading, setIsLoading] = useState<boolean>();
-  const [error, setError] = useState<string | undefined>();
-  const onDescriptionChange = useCallback(
-    (e: SyntheticEvent<HTMLTextAreaElement>) => {
-      setDescription(e.currentTarget.value);
-      setError(undefined);
-    },
-    []
-  );
-
-  const createIdea = useCallback(async () => {
-    try {
-      setIsLoading(true);
-      setError(undefined);
-      await mutation.mutateAsync({
-        description,
-      });
-      setIsLoading(false);
-      setDescription("");
-      onCreate();
-    } catch (err) {
-      setError("Your idea could not be saved. Try again later.");
-    }
-  }, [description, mutation, onCreate]);
-
-  return {
-    description,
-    error,
-    isLoading,
-    createIdea,
-    onDescriptionChange,
-  };
-}
+import useIdeaForm from "./useIdeasForm";
 
 type IdeaFormProps = {
   onCreate: () => void;
@@ -51,6 +11,7 @@ export default function IdeaForm({ onCreate }: IdeaFormProps) {
     useIdeaForm({
       onCreate,
     });
+
   const onFormSubmit = useCallback(
     (e: SyntheticEvent<HTMLFormElement>) => {
       e.preventDefault();
